finish

+++ src/app/heroes/heroes.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild, EventEmitter } from '@angular/core';
+import { MatTable } from '@angular/material/table';
 
 import { Hero } from '../hero';
 import { HeroService } from '../hero.service';
@@ -9,24 +10,21 @@ import { MessageService } from '../message.service';
 @Component({
   selector: 'app-heroes',
   templateUrl: './heroes.component.html',
-  styleUrls: ['./heroes.component.css']
+  styleUrls: ['./heroes.component.css'],
 })
 export class HeroesComponent implements OnInit {
+
+  @ViewChild(MatTable) table!: MatTable<Hero>;
+
+  displayedColumns: string[] = ['id', 'name', 'class', 'race', 'level', 'delete']
 
   heroes: Hero[] = [];
 
   classes: string[] = CLASSES;
   races: string[] = RACES;
 
-
   constructor(
-    private heroService: HeroService,
-    private memoryDataService: InMemoryDataService,
-    private message: MessageService) { }
-
-  private log(msg: string) {
-    this.message.add(`HeroesComponent: ${msg}`)
-  }
+    private heroService: HeroService,) {}
 
   ngOnInit(): void {
     this.getHeroes();
@@ -36,7 +34,7 @@ export class HeroesComponent implements OnInit {
     this.heroService.getHeroes()
     .subscribe(heroes => this.heroes = heroes);
   }
-
+/* 
   add(name: string, cls: string, rac: string, lvl: string) {
 
     if(!name || Number(lvl) < 0 || Number(lvl) > 100) {
@@ -53,13 +51,18 @@ export class HeroesComponent implements OnInit {
     };
 
     this.heroService.addHero(newHero)
-    .subscribe(hero => {
-      this.heroes.push(hero);
-    });
-  }
+      .subscribe(hero => {
+        this.heroes.push(hero);
+        this.table.renderRows();
+      });   
+    
+  } */
 
   delete(hero: Hero): void {
-    this.heroes = this.heroes.filter(h => h !== hero);
-    this.heroService.deleteHero(hero.id).subscribe();
+    const result = confirm('Do you realy want delete it?')
+    if(result) {
+      this.heroes = this.heroes.filter(h => h !== hero);
+      this.heroService.deleteHero(hero.id).subscribe();
+    }
   }
 }
